Upgrade to arrow 40 which has stronger/better schema safety
Upon upgrade to arrow 40, a schema mismatch error started to crop up to cause a
failure in the dv_table test. The underlying problem is/was that the remove
schema was incorrectly specifying that a non-null `stats` column which simply
doesn't exist. This is likely just a copy/paste error from the `add` action
since the Protocol[1] does not specify this column.

Older versions of arrow-rs do not properly validate schema conformance for
fields masking this bug.

This change also manifests a :bug: with the API design of `LogFile.read`, which
currently strives to return a Iterator of simply `Item = RecordBatch` while
`ReaderBuilder` will yield an iterator of `Item = Result<RecordBatch,
ArrowError>`. The erasure of the Result type in the `arrow::json::Reader::read`
loop results in an infinite loop on _any_ `ArrowError yielded by the underlying
Reader.

[1] <https://github.com/delta-io/delta/blob/master/PROTOCOL.md#add-file-and-remove-file>

Sponsored-by: Databricks Inc.