Replace StorageClient with Arc<dyn ObjectStore> to utilize the object_store crate
ObjectStore is a very mature and well-tested interface for interacting with a
myriad of storage systems and provides a flexible interface that will allow
engine implementations to bring their own new storage clients.

A side-effect of adopting ObjectStore is that it forces the introduction of
async code into the kernel, which has necessitated a *lot* more changes than
originally planned.

As such `impl Iterator` becomes `impl Stream` in all cases, and kernel adopts
the `futures` crate as a dependency. This helps kernel avoid incorporating a
specific runtime and instead will enable implementations on top to bring in
tokio, async-std, smol, etc as their needs dictate.

Sponsored-by: Databricks Inc