Convert delta-kernel-rs to a workspace project with multiple packages within it.
This structure builds on the work already done by @roeap which creates some
better acceptance testing infrastructure for the project. In doing so, the
project also gains a stronger acceptance test boundary because the acceptance
tests are in a completely separate project, making dependency requirements and
public APIs much cleaner to understand.

In order to make this change I did stumble across some unclean APIs that will
need to be cleaned up a little bit in future iterations

Sponsored-by: Databrics Inc