Remove unnecessary cloning of commit_files in the LogSegment
The commit_files() are _allegedly_ already sorted and none of the tests pass.
It's worth removing the cloning to keep operating on the slice as it was passed
in.

If LogSegment needs to do a mutation on `commit-files` then it will likely make
more sense to pull a `&mut self` up through the `replay()` function signature
through to the callers rather than clone. At which point it's worth considering
why the LogSegment should be mutable :smile:

Sponsored-by: Databricks Inc