Index the partitionValues map with column's physical name (#278)
From the protocol:
> In name mode, readers must resolve columns in the data files by their
physical names as given by the column metadata property
delta.columnMapping.physicalName in the Delta schema. Partition values
and column level statistics will also be resolved by their physical
names. For columns that are not found in the files, nulls need to be
returned. Column ids are not used in this mode for resolution purposes.

So we need to map to physical name otherwise we won't find the column.
Note the same is true for `id` mode, so this won't break when we add
support for that.