Update to dat 0.3 (#290)
- Update to the latest version of DAT
- Don't skip any tests

This requires some slight massaging of the expected data. In particular,
some of the expected timestamps are in zone `+00:00` which _is_ UTC
([link](https://en.wikipedia.org/wiki/UTC%2B00:00)), so if we find that,
we just convert.

This requires going column-by-column and normalizing if needed. 

This also removes the data_type check, since that actually needs to
happen after column normalization, and it does already happen as part of
checking if two columns are equal. See
[here](https://arrow.apache.org/rust/arrow_data/equal/fn.equal.html).