fix: allow kernel to read tables with invalid _last_checkpoint (#311)
Previously, `read_last_checkpoint` would return `Err(<json error>)` if
json parsing of the `_last_checkpoint` file failed. This caused the
upstream call to fail due to propagation of the error. Instead we now
return `None` from `read_last_checkpoint` in this case - similar to what
is returned if the `_last_checkpoint` is not found. This allows log
replay to continue without the hint and read the table.

Additionally, this unblocks the `corrupted-last-checkpoint-kernel`
golden table test.