Adding cargo semver checks (#369)
Adding a check that will run `cargo semver-checks` against each PR. 

This implements a check for PRs that don't bump the crate version. The
check will be run between the `HEAD` of the PR and `main`. So any
changes introduced in the PR vs main that break semver are picked up,
and the PR will be labeled with `breaking-change`.

Currently can't test the actual adding of a label because (probably) the
check needs to be in the main repo, rather than in a PR branch, to be
allowed to get write access to the repo.

So sadly to actually test this we'll need to merge and see if it works.
(hooray github actions dev workflow)