apply a schema to fix column names (#331)
Enforce/apply the schema given when we evaluate an expression. 

Given we want to go to expression based fixup and allow the final schema
to dictate the output, we will need to do this.

This code will fix-up at all levels of the output, which is messy in
arrow since schemas are embedded all over the place. The schema is only
applied if the output of the expression doesn't exactly match the passed
schema.

---------

Co-authored-by: Nick Lanham <nick.lanham@pm.me>
Co-authored-by: Ryan Johnson <scovich@users.noreply.github.com>