Expression::Column references a ColumnName object instead of a String (#400)
Kernel only partly supports nested column names today. One of the
biggest barriers to closing the gap is that Expression tracks column
names as simple strings, so there is no reliable way to deal with
nesting.

This PR takes a first step of defining a proper `ColumnName` struct for
hosting column names. So far, it behaves the same as before --
internally it's just a simple string, occasionally interpreted as nested
by splitting at periods. Changing the code to use this new construct is
noisy, so we make the changes as a pre-factor. A follow-up PR will
actually add nesting support.

Resolves https://github.com/delta-incubator/delta-kernel-rs/issues/422