Simplify log replay visitor and avoid materializing Add/Remove actions (#494)
## What changes are proposed in this pull request?

The existing log replay logic materialized full Add and Remove actions
in order to examine the four fields of each that comprise the file
comparison key. There was also a lot of indirection and duplication
because of an old (now-unused) flavor of log replay. It also created a
new expression evaluator for each batch instead of reusing it for the
whole iteration.

Streamline the logic to only visit the columns of interest and generally
reduce code bloat.

## How was this change tested?

Existing log replay unit tests.