Engines now return FileMeta with correct millisecond timestamps (#565)
## What changes are proposed in this pull request?
This PR fixes the `list_from` method used by both the `DefaultEngine`
and the `SyncEngine` to return the correct `FileMeta` struct.
Previously, both the engines would return the timestamp in number of
seconds from Unix Epoch. However, `FileMeta` specifies that we should
expect the modification time to be in milliseconds since Unix Epoch:
```
pub struct FileMeta {
    /// The fully qualified path to the object
    pub location: Url,
    /// The last modified time as milliseconds since unix epoch
    pub last_modified: i64,
    /// The size in bytes of the object
    pub size: usize,
}
```
### This PR affects the following public APIs
This changes the behaviour of the following:
- `SyncFilesystemClient::list_from`
- `ObjectStoreFileSystemClient::list_from`
- any transitive users of these methods


## How was this change tested?
I compare the `FileMeta` returned by each of the engines to expected
`FileMeta` within 10 seconds of creation time.